rename test cases
