+++ go.mod
@@ -1,0 +1,5 @@
+module github.com/mineiros-io/terraform-github-team/v2
+
+go 1.13
+
+require github.com/gruntwork-io/terratest v0.23.0 // indirect
